ai-web-scraper/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
lxml>=4.9.0